requirements.txt
alembic==1.4.2
Babel==2.8.0
blinker==1.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
dnspython==1.16.0
dominate==2.5.1
easydict==1.9
email-validator==1.1.1
eventlet==0.25.2
Flask==1.1.2
Flask-Avatar==0.1.3
Flask-Avatars==0.2.2
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-DotEnv==0.1.2
Flask-HTTPAuth==4.0.0
Flask-JSON==0.3.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.12.0
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-MQTT==1.1.0
flask-serialize==1.1.9
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.1
Flask-Threads==0.0.1a0
Flask-WTF==0.14.3
greenlet==0.4.15
httpie==2.1.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
login==0.0.6
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.6.0
marshmallow-sqlalchemy==0.23.1
monotonic==1.5
paho-mqtt==1.5.0
Pillow==7.1.2
psycopg2==2.8.5
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-engineio==3.12.1
python-socketio==4.5.1
pytz==2020.1
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
WTForms-JSON==0.3.3
